Update to the latest Maven plugin versions
Also update the minimum enforced Maven version for the project build to 3.3.1. This was already the minimum version supported at runtime, and the minimum version actually used to build in GitHub actions.

The maven-compiler-plugin has a new version that was not included in this update (3.13.0), because it requires Maven 3.6.3 or later.